More on configuration
Reached a place where more thought is required. The configparser mechanism is in place but it’s not sure how this should go forward.

Remove solve_image since it is redundant with the pipeline version

Wprojection now triggered by setting wstep > 0.0 in config

Removed legacy imaging definitions. Use the versions in imaging_context